# swagger/swagger.json
{"swagger":"2.0","info":{"title":"sample","description":"適当に作るよ","contact":{"name":"hoge","email":"hoge@hogehoge.com","url":"https://sample"},"license":{"name":"MIT","url":"https://hogehoge/license"},"version":""},"host":"localhost;8080","schemes":["http","https"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/api/v1/actions/hello":{"get":{"tags":["actions"],"summary":"hello actions","description":"挨拶します","operationId":"actions#hello","produces":["application/vnd.goa.error","text/plain"],"parameters":[{"name":"name","in":"query","description":"名前","required":true,"type":"string","default":""}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http","https"]}},"/api/v1/actions/ping":{"get":{"tags":["actions"],"summary":"ping actions","description":"サーバーとの疎通確認","operationId":"actions#ping","produces":["application/vnd.goa.error","text/plain"],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http","https"]}},"/api/v1/actions/{id}":{"get":{"tags":["actions"],"summary":"ID actions","description":"複数アクション(:id)","operationId":"actions#ID","produces":["application/vnd.goa.error","text/plain"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http","https"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http","https"]}},"/swagger/{filepath}":{"get":{"summary":"Download public/swagger/","operationId":"swagger#/swagger/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http","https"]}}},"definitions":{"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"NotFound":{"description":"Not Found"},"OK":{"description":"OK"}},"externalDocs":{"description":"wiki","url":"https://hoge/wiki"}}